Allow segments to be scoped to specified options
Allows a segment to be defined with a preset list of options. This may
be useful when targeting a small, known set of values for specific
segments (device types, country codes, etc.)

The initial plan was to allow this to be specified in the decorator
(`@segment(options=[...])`), however this is currently not possible
without a larger refactor of our approach to defining these decorators.
This implementation works fundamentally the same way, but with options
defined as a class level constant.